Fix race conditions
